requirements.txt
eventlet==0.36.1
Flask==2.2.5
Flask_Cors==4.0.0
flask_session==0.8.0
Flask_SocketIO==5.3.6
python-socketio==5.11.2
pytz==2023.3.post1